+++ K8sUtils/public/Invoke-HelmUpgrade.ps1
@@ -170,7 +170,7 @@ function Invoke-HelmUpgrade {
                     Write-Status "helm rollback failed, trying uninstall" -LogLevel Error -Char '-'
                     helm uninstall $ReleaseName | Out-File $OutputFile -Append
                 }
-                Write-Footer "End rolling back release '$ReleaseName' due to errors" -LogLevel Error
+                Write-Footer "End rolling back release '$ReleaseName' due to errors"
                 Remove-Item $errFile -ErrorAction SilentlyContinue
                 # throw "$msg, rolled back"
                 Write-Warning "$msg, rolled back"
@@ -262,7 +262,7 @@ function Invoke-HelmUpgrade {
         if ($DryRun) {
             Write-Status "Doing a helm dry run. Helm output and manifests follow."
         } else {
-            Write-Status ">> Helm upgrade$hookMsg"
+            Write-Header -Msg "Helm upgrade$hookMsg" -HeaderPrefix ""
         }
         # Helm's default timeout is 5 minutes. This doesn't return until preHook is done
         helm upgrade --install $ReleaseName $Chart -f $ValueFile --reset-values --timeout "${PreHookTimeoutSecs}s" --namespace $Namespace @parms 2>&1 | Tee-Object $tempFile -Append | Write-MyHost
@@ -271,11 +271,10 @@ function Invoke-HelmUpgrade {
         if ($DryRun) {
             return
         } elseif ($upgradeExit -eq 0) {
-            Write-Status "<< End Helm upgrade OK. (exit code $upgradeExit)"
+            Write-Footer "End Helm upgrade OK. (exit code $upgradeExit)" -FooterPrefix ""
         } else {
-            Write-Status "<< 👆 Check Helm output for error message 👆"
-            Write-Status "<< helm upgrade exited with: $upgradeExit"
-            Write-Status "<<"
+            Write-Footer "helm upgrade exited with: $upgradeExit" -FooterPrefix ""
+            Write-Status "👆 Check Helm output for error message 👆" -LogLevel Error
         }
 
         $hookStatus = $null